requirements.txt
streamlit
textblob
Pillow
PyMuPDF
python-docx
numpy
torch
sentence-transformers
scikit-learn
spacy
ollama
